restructuring; bugfixes
+restructured code a lot, made some new functions
+(hopefully) fixed most of the bugs;
+minor style changes, switched to google icons
