+++ README.md
@@ -2,3 +2,7 @@
 
 
 ### Currently in a very early stage of development. Available at https://miktaew.github.io/yet-another-idle-rpg/
+
+
+Expect complete lack of gameplay balance and a lot of bugs, some potentialy gamebreaking.  
+Using the "export" feature every now and then is highly recommended.
